created statemachine start moved MasqIntegrator.java to /MasqMath
