+++ Masquerade4997Library/src/main/java/Library4997/MasqControlSystems/MasqStateMachine/MasqState.java
@@ -1,0 +1,8 @@
+package Library4997.MasqControlSystems.MasqStateMachine;
+
+/**
+ * Created by Archishmaan Peyyety on 2019-11-14.
+ * Project: MasqLib
+ */
+public class MasqState {
+}
