--- Masquerade4997Library/src/main/java/Library4997/MasqMotors/MasqMotorV2.java
@@ -1,8 +1,0 @@
-package Library4997.MasqMotors;
-
-/**
- * Created by Archishmaan Peyyety on 4/12/19.
- * Project: MasqLib
- */
-public class MasqMotorV2 {
-}
